Moving functionalities around for ImageSetLazy and FormatMultiImageLazy. This avoids code duplication. The get_imageset function in FormatMultiImageLazy uses the flag lazy=True.
