+++ dxtbx/imageset.py
@@ -148,6 +148,61 @@ class ImageSetAux(boost.python.injector, ImageSet):
     return [self.get_path(i) for i in range(len(self))]
 
 
+class ImageSetLazy(ImageSet):
+
+  '''
+  Lazy ImageSet class that doesn't necessitate setting the models ahead of time.
+  Only when a particular model (like detector or beam) for an image is requested,
+  it sets the model using the format class and then returns the model
+  '''
+
+  def get_detector(self, index=None):
+    if index is None: index=0
+    detector = super(ImageSetLazy,self).get_detector(index)
+    if detector is None:
+      format_instance = self.get_format_class()._current_instance_
+      detector = format_instance.get_detector(self.indices()[index])
+      self.set_detector(detector,index)
+    return detector
+
+
+  def get_beam(self, index=None):
+    if index is None: index=0
+    beam = super(ImageSetLazy,self).get_beam(index)
+    if beam is None:
+      format_instance = self.get_format_class()._current_instance_
+      beam = format_instance.get_beam(self.indices()[index])
+      self.set_beam(beam,index)
+    return beam
+
+  def get_goniometer(self, index=None):
+    if index is None: index=0
+    goniometer = super(ImageSetLazy,self).get_goniometer(index)
+    if goniometer is None:
+      format_instance = self.get_format_class()._current_instance_
+      goniometer = format_instance.get_goniometer(self.indices()[index])
+      self.set_goniometer(goniometer,index)
+    return goniometer
+
+  def get_scan(self, index=None):
+    if index is None: index=0
+    scan = super(ImageSetLazy,self).get_scan(index)
+    if scan is None:
+      format_instance = self.get_format_class()._current_instance_
+      scan = format_instance.get_scan(self.indices()[index])
+      self.set_scan(scan,index)
+    return scan
+
+  def __getitem__(self, item):
+    if isinstance(item, slice):
+      return ImageSetLazy(self.data(), indices = self.indices()[item])
+    else:
+      # Sets the list for detector, beam etc before being accessed by functions in imageset.h
+      self.get_detector(item)
+      self.get_beam(item)
+      self.get_goniometer(item)
+      self.get_scan(item)
+    return super(ImageSetLazy,self).__getitem__(item)
 
 class ImageSweepAux(boost.python.injector, ImageSweep):
